requirements.txt
# Core Packages
Django==3.2.8
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-configurations==2.4
django-constance==2.9.1
django-dotenv==1.4.2
django-dynamic-fixture==3.1.2
django-extensions==3.2.1
django-model-utils==4.3.1
django-postgres-extra==2.0.8
django-url-filter==0.3.15
django-filter==2.2.0
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
django-picklefield==3.1
django-redis==4.11.0
drf-yasg==1.21.5
psycopg2==2.9.5

# Async Tasks
celery[redis]==5.2.3

# Utilities
coverage==7.1.0
cryptography==39.0.1
gunicorn==20.1.0
python-dotenv==0.21.1
pytz==2022.7.1
retrying==1.3.4
xlrd==2.0.1
tenacity==8.2.1
confluent-kafka==2.0.2
